+++ ja2lib/Strategic/StrategicTownLoyalty.c
@@ -397,7 +397,7 @@ void HandleMurderOfCivilian(struct SOLDIERTYPE *pSoldier, BOOLEAN fIntentional) 
     struct SOLDIERTYPE *pKiller = MercPtrs[pSoldier->ubAttackerID];
 
     // apply morale penalty for killing a civilian!
-    HandleMoraleEvent(pKiller, MORALE_KILLED_CIVILIAN, pKiller->sSectorX, pKiller->sSectorY,
+    HandleMoraleEvent(pKiller, MORALE_KILLED_CIVILIAN, (u8)pKiller->sSectorX, (u8)pKiller->sSectorY,
                       pKiller->bSectorZ);
   }
